+++ packages/lib/src/model/Model.ts
@@ -20,16 +20,17 @@ export type _ComposedCreationData<
   ? ModelPropsToTransformedCreationData<TProps> & TCD
   : ModelPropsToTransformedCreationData<TProps>
 
+export type _FromSnapshot<TProps extends ModelProps> = SnapshotInOfObject<
+  ModelPropsToCreationData<TProps>
+>
+
+export type _ToSnapshot<TProps extends ModelProps> = SnapshotOutOfObject<ModelPropsToData<TProps>>
+
 export type _ModelId<SuperModel, TProps extends ModelProps> = SuperModel extends AnyModel
   ? ModelIdPropertyName<SuperModel>
   : ExtractModelIdProp<TProps> & string
 
-export interface _Model<
-  SuperModel,
-  TProps extends ModelProps,
-  FromSnapshot = SnapshotInOfObject<ModelPropsToCreationData<TProps>>,
-  ToSnapshot = SnapshotOutOfObject<ModelPropsToData<TProps>>
-> {
+export interface _Model<SuperModel, TProps extends ModelProps, FromSnapshot, ToSnapshot> {
   new (data: _ComposedCreationData<SuperModel, TProps>): SuperModel &
     BaseModel<
       ModelPropsToData<TProps>,
@@ -49,47 +50,6 @@ export interface _Model<
 }
 
 /**
- * Optional transformation that will be run when converting from a snapshot to the data part of the model.
- * Useful for example to do versioning and keep the data part up to date with the latest version of the model.
- *
- * @param model The model to apply the transformation over.
- * @param fn The function that will convert the custom input snapshot. Returns an input snapshot that must
- * match the current model input snapshot.
- * @returns The model with the transform.
- */
-export function fromSnapshotProcessor<
-  FS,
-  SuperModel,
-  TProps extends ModelProps,
-  FromSnapshot,
-  ToSnapshot
->(
-  model: _Model<SuperModel, TProps, FromSnapshot, ToSnapshot>,
-  fn: (sn: FS) => FromSnapshot
-): _Model<SuperModel, TProps, FS, ToSnapshot> {
-  return (model as any).withFromSnapshotProcessor(fn)
-}
-
-/**
- * Optional transformation that will be run when converting the data part of the model into a snapshot.
- *
- * @param fn The function that will convert the output snapshot. Returns a custom output snapshot.
- * @returns  The model with the transform.
- */
-export function toSnapshotProcessor<
-  TS,
-  SuperModel,
-  TProps extends ModelProps,
-  FromSnapshot,
-  ToSnapshot
->(
-  model: _Model<SuperModel, TProps, FromSnapshot, ToSnapshot>,
-  fn: (sn: ToSnapshot, modelInstance: any) => TS
-): _Model<SuperModel, TProps, FromSnapshot, TS> {
-  return (model as any).withToSnapshotProcessor(fn)
-}
-
-/**
  * Extract the model id property from the model props.
  */
 export type ExtractModelIdProp<TProps extends ModelProps> = {
@@ -105,13 +65,19 @@ export type ExtractModelIdProp<TProps extends ModelProps> = {
  * @param modelOptions Model options.
  * @returns
  */
-export function ExtendedModel<TProps extends ModelProps, TModel extends AnyModel, A extends []>(
+export function ExtendedModel<
+  TProps extends ModelProps,
+  TModel extends AnyModel,
+  A extends [],
+  FS = _FromSnapshot<TProps>,
+  TS = _ToSnapshot<TProps>
+>(
   genFn: (...args: A) => {
     baseModel: AbstractModelClass<TModel>
     props: TProps
   },
-  modelOptions?: ModelOptions
-): _Model<TModel, TProps>
+  modelOptions?: ModelOptions<TProps, FS, TS>
+): _Model<TModel, TProps, FS, TS>
 
 /**
  * Base abstract class for models that extends another model.
@@ -123,16 +89,24 @@ export function ExtendedModel<TProps extends ModelProps, TModel extends AnyModel
  * @param modelOptions Model options.
  * @returns
  */
-export function ExtendedModel<TProps extends ModelProps, TModel extends AnyModel>(
+export function ExtendedModel<
+  TProps extends ModelProps,
+  TModel extends AnyModel,
+  FS = _FromSnapshot<TProps>,
+  TS = _ToSnapshot<TProps>
+>(
   baseModel: AbstractModelClass<TModel>,
   modelProps: TProps,
-  modelOptions?: ModelOptions
-): _Model<TModel, TProps>
+  modelOptions?: ModelOptions<TProps, FS, TS>
+): _Model<TModel, TProps, FS, TS>
 
 // base
-export function ExtendedModel<TProps extends ModelProps, TModel extends AnyModel>(
-  ...args: any[]
-): _Model<TModel, TProps> {
+export function ExtendedModel<
+  TProps extends ModelProps,
+  TModel extends AnyModel,
+  FS = _FromSnapshot<TProps>,
+  TS = _ToSnapshot<TProps>
+>(...args: any[]): _Model<TModel, TProps, FS, TS> {
   let baseModel
   let modelProps
   let modelOptions
@@ -162,10 +136,15 @@ export function ExtendedModel<TProps extends ModelProps, TModel extends AnyModel
  * @param fnModelProps Function that generates model properties.
  * @param modelOptions Model options.
  */
-export function Model<TProps extends ModelProps, A extends []>(
+export function Model<
+  TProps extends ModelProps,
+  A extends [],
+  FS = _FromSnapshot<TProps>,
+  TS = _ToSnapshot<TProps>
+>(
   fnModelProps: (...args: A) => TProps,
-  modelOptions?: ModelOptions
-): _Model<unknown, TProps>
+  modelOptions?: ModelOptions<TProps, FS, TS>
+): _Model<unknown, TProps, FS, TS>
 
 /**
  * Base abstract class for models.
@@ -176,16 +155,21 @@ export function Model<TProps extends ModelProps, A extends []>(
  * @param modelProps Model properties.
  * @param modelOptions Model options.
  */
-export function Model<TProps extends ModelProps>(
-  modelProps: TProps,
-  modelOptions?: ModelOptions
-): _Model<unknown, TProps>
+export function Model<
+  TProps extends ModelProps,
+  FS = _FromSnapshot<TProps>,
+  TS = _ToSnapshot<TProps>
+>(modelProps: TProps, modelOptions?: ModelOptions<TProps, FS, TS>): _Model<unknown, TProps, FS, TS>
 
 // base
-export function Model<TProps extends ModelProps>(
+export function Model<
+  TProps extends ModelProps,
+  FS = _FromSnapshot<TProps>,
+  TS = _ToSnapshot<TProps>
+>(
   fnModelPropsOrModelProps: (() => TProps) | TProps,
-  modelOptions?: ModelOptions
-): _Model<unknown, TProps> {
+  modelOptions?: ModelOptions<TProps, FS, TS>
+): _Model<unknown, TProps, FS, TS> {
   const modelProps =
     typeof fnModelPropsOrModelProps === "function"
       ? fnModelPropsOrModelProps()
@@ -193,28 +177,51 @@ export function Model<TProps extends ModelProps>(
   return internalModel(modelProps, undefined, modelOptions)
 }
 
-function internalModel<TProps extends ModelProps, TBaseModel extends AnyModel>(
+function internalModel<
+  TProps extends ModelProps,
+  TBaseModel extends AnyModel,
+  FS = _FromSnapshot<TProps>,
+  TS = _ToSnapshot<TProps>
+>(
   modelProps: TProps,
   baseModel: ModelClass<TBaseModel> | undefined,
-  modelOptions?: ModelOptions
-): _Model<TBaseModel, TProps> {
+  modelOptions?: ModelOptions<TProps, FS, TS>
+): _Model<TBaseModel, TProps, FS, TS> {
   return sharedInternalModel({
     modelProps,
     baseModel,
     type: "class",
     valueType: modelOptions?.valueType ?? false,
-    fromSnapshotProcessor: undefined,
-    toSnapshotProcessor: undefined,
+    fromSnapshotProcessor: modelOptions?.fromSnapshotProcessor,
+    toSnapshotProcessor: modelOptions?.toSnapshotProcessor,
   })
 }
 
 /**
  * Model options.
  */
-export interface ModelOptions {
+export interface ModelOptions<TProps extends ModelProps, FS, TS> {
   /**
    * A value type will be cloned automatically when being attached to a new tree.
    * The default is `false`.
    */
   valueType?: boolean
+
+  /**
+   * Optional transformation that will be run when converting from a snapshot to the data part of the model.
+   * Useful for example to do versioning and keep the data part up to date with the latest version of the model.
+   *
+   * @param sn The custom input snapshot.
+   * @returns An input snapshot that must match the expected model input snapshot.
+   */
+  fromSnapshotProcessor?(sn: FS): _FromSnapshot<TProps>
+
+  /**
+   * Optional transformation that will be run when converting the data part of the model into a snapshot.
+   *
+   * @param sn The output snapshot.
+   * @param modelInstance The model instance.
+   * @returns  a custom output snapshot.
+   */
+  toSnapshotProcessor?(sn: _ToSnapshot<TProps>, modelInstance: any): TS
 }